test_project/requirements.txt
django-ace>=1.0.1
django-classy-tags>=0.4
django-haystack>=2.0,<3
django-imagekit>=3.0.4
django-model-utils>=2.0
django-tinymce>=1.5.2
django-wymeditor>=1.0
-e git+https://github.com/kezabelle/django-adminlinks.git#egg=django-adminlinks
feedparser>=5.1.3,<6.0
Pillow>=2.3
importd>=0.2.9
Whoosh>=2.5.6
